Update to move ResourceDowmloadQueue into its own file
